{"schema":"decodex/commit/1","summary":"Split MCP HTTP modules","authority":"XY-1115"}
